+++ Source/Obsidian/Private/AbilitySystem/ModMagCalculations/ObsidianMMC_MaxHealth.cpp
@@ -50,7 +50,7 @@ float UObsidianMMC_MaxHealth::CalculateBaseMagnitude_Implementation(const FGamep
 		return Super::CalculateBaseMagnitude_Implementation(Spec);
 	}
 	
-	const int32 CharacterLevel = CombatInterface->GetCharacterLevel();
+	const uint8 CharacterLevel = CombatInterface->GetCharacterLevel();
 	
 	const FAggregatorEvaluateParameters EvaluationParameters;
