+++ Source/Obsidian/Private/AbilitySystem/ModMagCalculations/ObsidianMMC_MaxMana.cpp
@@ -48,7 +48,7 @@ float UObsidianMMC_MaxMana::CalculateBaseMagnitude_Implementation(const FGamepla
 		return Super::CalculateBaseMagnitude_Implementation(Spec);
 	}
 	
-	const int32 CharacterLevel = CombatInterface->GetCharacterLevel();
+	const uint8 CharacterLevel = CombatInterface->GetCharacterLevel();
 	
 	const FAggregatorEvaluateParameters EvaluationParameters;
