+++ Source/Obsidian/Public/UI/CharacterStatus/ObsidianCharacterStatus.h
@@ -42,7 +42,7 @@ protected:
 	void SetExperienceTextBlock() const;
 	void SetExperienceProgressBar() const;
 
-	void OnHeroLevelUp(const int32 NewLevel);
+	void OnHeroLevelUp(const uint8 NewLevel);
 	
 	void OnExperienceChanged(const float Value);
 	void OnMaxExperienceChanged(const float Value, const float OldValue);
